Refactor: folder 명 수정
